\name{fruitfly}
\docType{data}
\alias{fruitfly}
\title{Drosophila microarray time course data in Tomancak et al. (2002)}
\description{
This is a subset of the Drosophila microarray time course data in
Tomancak et al. (2002). 
}
\usage{data(fruitfly)}
\format{A matrix of \eqn{log_2} gene expression values for 2000 probesets.}
\source{
Tomancak et al. (2002) \emph{Systematic determination of patterns of gene expression during 
Drosophila embryogenesis.}
Genome Biology 2002, 3:research0088.1-0088.14
\url{http://genomebiology.com/2002/3/12/research/0088.1}
The complete dataset can be downloaded from the following website
\url{http://www.fruitfly.org/cgi-bin/ex/insitu.pl}
}

\keyword{datasets}
