\name{ot.helmert}
\alias{ot.helmert}

\title{Helmert orthogonal transformation}
\description{Computes the Helmert orthogonal transformation matrix.}
\usage{
ot.helmert(k)
}
\arguments{
  \item{k}{a positive integer giving the number of time points.}         
}
\value{
a numeric matrix.
}

\details{
This function is for internal use only and is not to be called by the user.
}


\author{Yu Chuan Tai \email{yuchuan@stat.berkeley.edu}}
\keyword{array}


